requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
dj-database-url==2.3.0
Django==5.2.1
django-unfold==0.57.0
environs==8.0.0
Faker==37.1.0
frozenlist==1.6.0
idna==3.10
magic-filter==1.0.12
marshmallow==3.20.2
multidict==6.4.3
packaging==25.0
pillow==11.2.1
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
setuptools==80.4.0
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
yarl==1.20.0
gunicorn==23.0.0

requests~=2.32.4
drf-yasg~=1.21.10
djangorestframework~=3.16.0